fix(tabs): rename Tabs.type in order to make more accurate its use
