+++ src/components/navigation/Tabs/types/Tabs.type.ts
